删掉 runtime_introspection：它是同一批数据的第二套读法，而且更弱
结论是实测出来的，不是看代码猜的
--------------------------------
它看起来像"设计完整但没人接"的模块：4 个公开导出、24 KB、一整套测试。查下来
不是这么回事。

首先它和那些手拼的 introspection_snapshot **不是生产者/消费者关系** —— 抽取器
根本不读 introspection_snapshot 这个键，两边各自去读同一批结果字典的原始字段。
是两套并行读法。

然后拿**真实的 route_envelope 结果**喂进去：

  RuntimeIntrospectionSnapshot 抽到      4 / 23 个有意义字段
  这 4 个                                execution_substrate_role / lifecycle_state
                                         / execution_mode / request_success
  同一结果自带的手写 dict 有没有          全都有
  手写 dict 多出、它给不出的              execution_path / failure_is_retryable
                                         / tool_invocation_truth / repo_mutation_truth

其中 execution_path 是唯一有真实生产读取方的字段（command_router:3006 拿它决定
要不要记本地执行链）。也就是说在最要紧的那一层，它比现有的 10 行 dict 更弱。

为什么弱：抽取器读 metadata / authority_metadata 子字典。openclawd 的结果有
（authority_role、delegation_point、kernel_cognition_role 都在），command_router
的结果是扁的、没有 metadata，于是抽了个空。它在有些层能用，在另一些层近乎全空。

而它真正的用法是跨三层聚合 build_introspection_snapshot(shell, subject_core,
substrate)。全仓**没有任何一处同时握着这三份结果** —— 每层只把扁平结果返给上层，
不留下层的结果字典（openclawd 构 TaskEnvelope 调 route_envelope，取个值就丢）。
所以聚合模式没有可能的调用方，除非改三个大模块返回什么。

那个想法（跨层解释一次请求走了什么路）本身有价值，仓里也没有别的东西干这件事。
死的是这一份实现拿不到数据。真要做，应该从"每层留不留下层结果"这个决定开始，
而不是从复活这个模块开始。

连带处理
--------
* tools/architecture/architecture_status_from_snapshot() 一并删除：它唯一的入参
  类型是 RuntimeIntrospectionSnapshot，输入来源没了。
  build_architecture_status_report() 吃各层原始 dict，不依赖它，保留。
* 测试文件从 test_pr14_runtime_introspection.py 更名为
  test_architecture_status_report.py，只留还测活代码的三组（16 passed）。
  顺带删掉 S 组——它只断言**本文件自己的 fixture** 里有 introspection_snapshot
  这个键，没有碰任何生产代码，是条自证的空测试。真正钉住该字段的是
  tests/integration/runtime/test_runtime_integration.py，跑的是真实 route_envelope。
* reachability / wiring 两个基线随之收敛。

顺带更正一处我自己的说法：上一轮收尾我说
core/cross_device_responsiveness_contract.py 也还没接完 —— 错的。它的入口
responsiveness_for_participant 有 2 处生产调用（选择侧 + 事件侧），__all__ 里
其余 12 项是经由那个入口用到的枚举/策略哨兵/输入类型，不是各自待接的能力。

净删 1699 行。全量 4 分片 40726 passed, 0 failed；lint / 复杂度 / 接线 / 可达性 /
卫生五个守卫与 main.py doctor 均通过。
