+++ config/wiring_baseline.json
@@ -304,7 +304,6 @@
     "install_dry_run",
     "interpret_replay_sequence",
     "interpret_snapshot",
-    "introspection_snapshot_from_response",
     "is_active_desktop_status_directory",
     "is_android_inbound_path",
     "is_android_main_chain_eligible",
@@ -445,7 +444,6 @@
     "mark_task_assigned",
     "mark_task_released",
     "mark_verified",
-    "merge_layer_snapshots",
     "merge_videos",
     "mine_full",
     "native_speech_backend_registered",
